spring-cloud-modules/spring-cloud-zuul-fallback/spring-cloud-zuul-fallback-api-gateway/src/main/java/cn/tuyucheng/taketoday/spring/cloud/apigateway/fallback/WeatherServiceFallback.java
package cn.tuyucheng.taketoday.spring.cloud.apigateway.fallback;

import com.netflix.hystrix.exception.HystrixTimeoutException;
import org.springframework.cloud.netflix.zuul.filters.route.FallbackProvider;
import org.springframework.http.HttpStatus;
import org.springframework.http.client.ClientHttpResponse;
import org.springframework.stereotype.Component;

@Component
class WeatherServiceFallback implements FallbackProvider {

   private static final String DEFAULT_MESSAGE = "Weather information is not available.";

   @Override
   public String getRoute() {
      return "weather-service";
   }

   @Override
   public ClientHttpResponse fallbackResponse(String route, Throwable cause) {
      if (cause instanceof HystrixTimeoutException) {
         return new GatewayClientResponse(HttpStatus.GATEWAY_TIMEOUT, DEFAULT_MESSAGE);
      } else {
         return new GatewayClientResponse(HttpStatus.INTERNAL_SERVER_ERROR, DEFAULT_MESSAGE);
      }
   }
}